Update first-instance.rst
Added a line of white space under every image block in Launch from Dashboard, so the text for each image sits more clearly with the image.

Also added a note to Access and Security subsection, advising users to add Port 80 (HTTP) and/or 443 (HTTPS) if they want to set up an internet webserver.  

Then edited the warning about using the CIDR 0.0.0.0/0, which appears there, so the note and warning are consistent.